move assignor test to assingor package
